refactor TodayScreen
