+++ src/screens/index.js
@@ -1,4 +1,4 @@
-export { default as TodayScreen } from './TodayScreen/TodayScreen'
+export { default as TodayScreen } from './TodayScreen'
 export { default as WeekScreen } from './WeekScreen/WeekScreen'
 export { default as ActivityDetailScreen} from './ActivityDetailScreen/ActivityDetailScreen'
 export { default as GoalsScreen } from './GoalsScreen/GoalsScreen'
